requirements.txt
# gestion-facturas - dependencias fijadas (01/03/2026)
# Instalar: pip install -r requirements.txt

# Core
pandas==2.3.0
openpyxl==3.1.5
requests==2.32.5
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

# WooCommerce API
WooCommerce==3.0.0

# Gmail API
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
google-api-python-client==2.187.0

# PDF extraction
pdfplumber==0.11.7
pytesseract==0.3.13
pdf2image==1.17.0

# Dashboard PDF
matplotlib==3.10.0
reportlab==4.4.0

# Imágenes (dependencia de pdf2image y reportlab)
Pillow==11.2.1

# Fuzzy matching (cuadre)
rapidfuzz==3.13.0

# API backend (Fase 0+1)
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
